Activo con Notificaciones + Seguridad + Usuarios
Lo mismo que con las demás, le arreglé unas cosas del lado del server que podían ahcer que se cayese

+++ Servidor/controllers/casoActivoController.js
@@ -1,10 +1,16 @@
 const casoActivoModel = require('../models/casoActivoModel')
 const casoExcluidoModel = require('../models/casoExcluidoModel')
+const auth = require('./authController')
 const mongoose = require('mongoose')
 const usuarioModel = require('../models/usuarioModel')
 const uuidv4 = require('uuid/v4');
 
 function getCasosActivos(req, res) {
+  if(req.query.token == "undefined" || !auth.autentificarAccion(req.query.token)){
+    res.status(100)
+    res.json({ error: true , casos: []})
+    return
+  }
   casoActivoModel.find()
     .exec((err, casos) => {
       if (err) {
@@ -17,34 +23,80 @@ function getCasosActivos(req, res) {
 }
 
 function createCasoActivo(req,res){
-  let newCaso = new casoActivoModel(req.body)
+  let usuario = req.body.usuario;
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
+  let newCaso = new casoActivoModel(req.body.caso)
+  let notificacion = { autor: usuario.usuario, _id: uuidv4(), fecha: new Date(), location: "activo", action: "create", caso: newCaso._id }
   newCaso.save((err, resp) => {
     if(err){
       res.status(500)
       res.send({error:true})
     }
+    else{
+      res.status(200)
+      res.send({error:false, caso:newCaso})
+      usuarioModel.updateMany({ "$push": { "notificaciones": notificacion } }).exec()
+    }
   })
-  res.status(200)
-  res.send({error:false, caso:newCaso})
 }
 
 function editCasoActivo(req, res) {
-  let notificacion = {autor:"kevin",_id:uuidv4(),fecha:new Date(),location:"activos",action:"update",caseId:req.body._id}
-  casoActivoModel.updateOne({_id: new mongoose.Types.ObjectId(req.body._id)}, {$set: req.body})
+  let usuario = req.body.usuario;
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
+
+  let notificacion = {autor:usuario.usuario,_id:uuidv4(),fecha:new Date(),location:"activo",action:"update",caso:{}}
+  
+  casoActivoModel.findByIdAndUpdate({_id: new mongoose.Types.ObjectId(req.body.caso._id)}, {$set: req.body.caso},{new:true})
     .exec((err, caso) => {
       if (err) {
         res.status(500)
-        res.send({error:false})
+        res.send({error:true})
       }
       else{
         res.status(200)
-        res.send({error:false,caso:{...req.body}})
+        res.send({error:false,caso:caso})
+        notificacion.caso=caso._id
         usuarioModel.updateMany({"$push": { "notificaciones": notificacion } }).exec()
       }
     })
 }
 
 function excludeCasoActivo(req, res) {
+  let usuario = req.body.usuario;
+
+  if(usuario.token===undefined){
+    res.status(500)
+    res.send({ error: true , type: 0})
+    return
+  }
+
+  if(!auth.autentificarAccion(usuario.token)){
+    res.status(500)
+    res.send({ error: true , type: 1})
+    return
+  }
   casoActivoModel.deleteOne({_id: new mongoose.Types.ObjectId(req.params.id)})
     .exec((err, caso) => {
       if (err) {
@@ -52,10 +104,11 @@ function excludeCasoActivo(req, res) {
         res.send(`Ocurrió un error 💩 ${err}`)
       }
       let newCaso = new casoExcluidoModel({cedula: req.body.caso.cedula, apellidos: req.body.caso.apellidos, 
-        nombre: req.body.caso.nombre, domicilio: req.body.caso.domicilio, señas: req.body.caso.señas, telefono: req.body.caso.telefono,
+        _id: new mongoose.Types.ObjectId(req.params.id),nombre: req.body.caso.nombre, domicilio: req.body.caso.domicilio, señas: req.body.caso.señas, telefono: req.body.caso.telefono,
         ingreso: req.body.caso.ingreso, inicio: req.body.caso.inicio, altv_aprobadas: req.body.caso.alternativas, nacimiento: req.body.caso.nacimiento,
         sede: req.body.caso.sede, notas:req.body.caso.notas, nacimiento: req.body.caso.nacimiento })
-      let notificacion = {autor:"kevin",_id:uuidv4(),fecha:new Date(),location:"activo",action:"accepted", caseId:newCaso._id}
+      
+        let notificacion = {autor:usuario.usuario,_id:uuidv4(),fecha:new Date(),location:"activo",action:"accepted", caso:newCaso._id}
       newCaso.save((err, resp) => {
         if(err){
           res.status(500)
